report sale upload

+++ src/components/MediaGrid.jsx
@@ -9,6 +9,7 @@ import { useEffect, useState } from "react";
 import { addPhotos } from "../redux/services/mediaSlice";
 import { useContextCustom } from "../context/stateContext";
 import MediaImgDetail from "./MediaImgDetail";
+import Swal from "sweetalert2";
 
 const MediaGrid = ({ imgs }) => {
   MediaGrid.propTypes = {
@@ -28,24 +29,21 @@ const MediaGrid = ({ imgs }) => {
   }, [imgs]);
 
   const deletePhotoHandler = async (id) => {
-    const { data } = await deletePhoto({ id, token });
-    console.log("del", data);
-
-    // Swal.fire({
-    //   title: "Are you sure?",
-    //   text: "You won't be able to revert this!",
-    //   icon: "warning",
-    //   showCancelButton: true,
-    //   confirmButtonColor: "#3085d6",
-    //   cancelButtonColor: "#d33",
-    //   confirmButtonText: "Yes, delete it!",
-    // }).then(async (result) => {
-    //   if (result.isConfirmed) {
-    //     Swal.fire("Deleted!", "Your file has been deleted.", "success");
-    //     const { data } = await deletePhoto({ id, token });
-    //     console.log('del',data);
-    //   }
-    // });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        Swal.fire("Deleted!", "Your file has been deleted.", "success");
+        const { data } = await deletePhoto({ id, token });
+        console.log('del',data);
+      }
+    });
   };
   const imgModalHandler = (index) => {
     setImgIndex(index);
